chore: split Caption into smaller components
