+++ packages/react-day-picker/src/components/CaptionDropdowns/index.ts
@@ -1,0 +1,1 @@
+export * from './CaptionDropdowns';
