+++ packages/react-day-picker/src/components/CaptionNavigation/index.ts
@@ -1,0 +1,1 @@
+export * from './CaptionNavigation';
